# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Woolwire API
  description: HTTP API specification for Woolwire local owner, room creator, bootstrap, and peer interfaces.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:7070
    description: Local owner interface
paths:
...
  /bootstrap/v1/join:
    post:
      summary: Submit join request to room creator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [room_id, invitation_id, admission_secret, device_public, display_name, tailcat_addr]
              properties:
                room_id:
                  type: string
                invitation_id:
                  type: string
                admission_secret:
                  type: string
                device_public:
                  type: string
                display_name:
                  type: string
                tailcat_addr:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [admitted, pending, rejected]
                  membership:
                    type: object
                  roster:
                    type: array
                    items:
                      type: object

  # Peer API (Peer transport listener)
  /peer/v1/membership/sync:
    post:
      summary: Exchange membership updates with a peer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                known_version:
                  type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  roster_version:
                    type: integer
                  members:
                    type: array
                    items:
                      type: object